first cut of bittorrent support and download bandwidth throttling
git-svn-id: svn://svn.berlios.de/gpodder/trunk@312 b0d088ad-0a06-0410-aad2-9ed5178a7e87

+++ src/gpodder/libgpodder.py
@@ -37,6 +37,7 @@ from xml.sax import make_parser
 from string import strip
 from os.path import expanduser
 from os.path import exists
+from os.path import splitext
 from liblogger import log
 try:
     from os.path import lexists
@@ -124,6 +125,10 @@ class gPodderLibClass( object):
         self.opml_url = ""
         self.update_on_startup = False
         self.download_after_update = False
+        self.torrentdir = expanduser('~/gpodder-downloads/torrents')
+        self.use_gnome_bittorrent = True
+        self.limit_rate = False
+        self.limit_rate_value = 4.0
         self.show_played = False
         self.update_tags = False
         self.desktop_link = _("gPodder downloads")
@@ -181,7 +186,7 @@ class gPodderLibClass( object):
         for ddir in download_dirs:
             if isdir( ddir):
                 globr = glob( '%s/*' % ( ddir, ))
-                if not globr:
+                if not globr and ddir != self.torrentdir:
                     log( 'Stale download directory found: %s', basename( ddir))
                     try:
                         rmdir( ddir)
@@ -198,10 +203,14 @@ class gPodderLibClass( object):
         self.write_to_parser( parser, 'ipod_mount', self.ipod_mount)
         self.write_to_parser( parser, 'update_on_startup', self.update_on_startup)
         self.write_to_parser( parser, 'download_after_update', self.download_after_update)
+        self.write_to_parser( parser, 'limit_rate', self.limit_rate)
+        self.write_to_parser( parser, 'limit_rate_value', self.limit_rate_value)
         self.write_to_parser( parser, 'show_played', self.show_played)
         self.write_to_parser( parser, 'update_tags', self.update_tags)
         self.write_to_parser( parser, 'opml_url', self.opml_url)
         self.write_to_parser( parser, 'download_dir', self.downloaddir)
+        self.write_to_parser( parser, 'bittorrent_dir', self.torrentdir)
+        self.write_to_parser( parser, 'use_gnome_bittorrent', self.use_gnome_bittorrent)
         self.write_to_parser( parser, 'device_type', self.device_type)
         self.write_to_parser( parser, 'main_window_width', self.main_window_width)
         self.write_to_parser( parser, 'main_window_height', self.main_window_height)
@@ -288,6 +297,13 @@ class gPodderLibClass( object):
         except:
             return default
 
+    def get_float_from_parser( self, parser, option, default = 1.0):
+        try:
+            result = float(parser.get( self.gpodderconf_section, option))
+            return result
+        except:
+            return default
+
     def get_boolean_from_parser( self, parser, option, default = False):
         try:
             result = parser.getboolean( self.gpodderconf_section, option)
@@ -327,9 +343,13 @@ class gPodderLibClass( object):
                     self.ipod_mount = self.get_from_parser( parser, 'ipod_mount', '/media/ipod')
                     self.update_on_startup = self.get_boolean_from_parser(parser, 'update_on_startup', default=False)
                     self.download_after_update = self.get_boolean_from_parser(parser, 'download_after_update', default=False)
+                    self.limit_rate = self.get_boolean_from_parser(parser, 'limit_rate', default=False)
+                    self.limit_rate_value = self.get_float_from_parser(parser, 'limit_rate_value', default=4.0)
                     self.show_played = self.get_boolean_from_parser(parser, 'show_played', default=False)
                     self.update_tags = self.get_boolean_from_parser(parser, 'update_tags', default=False)
                     self.downloaddir = self.get_from_parser( parser, 'download_dir', expanduser('~/gpodder-downloads'))
+                    self.torrentdir = self.get_from_parser( parser, 'bittorrent_dir', expanduser('~/gpodder-downloads/torrents'))
+                    self.use_gnome_bittorrent = self.get_boolean_from_parser( parser, 'use_gnome_bittorrent', default=True)
                     self.device_type = self.get_from_parser( parser, 'device_type', 'none')
                     self.main_window_width = self.get_int_from_parser( parser, 'main_window_width', 600)
                     self.main_window_height = self.get_int_from_parser( parser, 'main_window_height', 450)
@@ -361,6 +381,8 @@ class gPodderLibClass( object):
             self.mp3_player_folder = '/media/usbdisk'
             self.opml_url = default_opml_directory
             self.downloaddir = expanduser('~/gpodder-downloads')
+            self.torrentdir = expanduser('~/gpodder-downloads/torrents')
+            self.use_gnome_bittorrent = True
         if was_oldstyle:
             self.saveConfig()
 
@@ -473,6 +495,21 @@ class gPodderLibClass( object):
         except:
             # silently ignore 
             pass
+            
+    def invoke_torrent( self, url, torrent_filename, target_filename):
+        self.history_mark_played( url)
+
+        if self.use_gnome_bittorrent:
+            command = 'gnome-btdownload "%s" --saveas "%s"' % ( torrent_filename, join( self.torrentdir, target_filename))
+            log( command, sender = self)
+            system( '%s &' % command)
+        else:
+            # Simply copy the .torrent with a suitable name
+            try:
+                target_filename = join( self.torrentdir, splitext( target_filename)[0] + '.torrent')
+                shutil.copyfile( torrent_filename, target_filename)
+            except:
+                log( 'Torrent copy failed: %s => %s.', torrent_filename, target_filename)
 
 class gPodderChannelWriter( object):
     def write( self, channels):
